[Android] New usage for AppHooks
Chromium change:
https://github.com/chromium/chromium/commit/530c14a0f529f79f84a536f60c18b1f7c7a6613e

[PolicyAuditor] Decouple PolicyAuditor from java WebContentsObservers
During navigations, we call down into java to notify an observer which
checks if we have a PolicyAuditor. Most of the time, we don't have
a PolicyAuditor, so these calls are unnecessary. This cl moves this
check to native, so in the case where we don't have a PolicyAuditor, we
don't create the observer.

Bug: 1353905